+++ release_notes.md
@@ -3,3 +3,4 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
+- Improved memory metrics reporting using CGroup data for Linux consumption (#10968)